25/09/2022

+++ orders/models.py
@@ -63,7 +63,7 @@ class OrderProduct(models.Model):
     product         = models.ForeignKey(Products,on_delete=models.CASCADE)
     variation       = models.ManyToManyField(Variation,blank=True)
     quantity        = models.IntegerField()
-    product_price   = models.FloatField()
+    product_price   = models.IntegerField()
     ordered         = models.BooleanField(default=False)
     created_at      = models.DateTimeField(auto_now_add=True)
     updated_at      = models.DateTimeField(auto_now=True)
